+++ md2cf/document.py
@@ -56,6 +56,7 @@ def get_pages_from_directory(
     collapse_empty: bool = False,
     beautify_folders: bool = False,
     use_pages_file: bool = False,
+    strip_header: bool = False,
     use_gitignore: bool = True
 ) -> List[Page]:
     """
@@ -139,7 +140,9 @@ def get_pages_from_directory(
                 )
 
         for markdown_file in markdown_files:
-            processed_page = get_page_data_from_file_path(markdown_file)
+            processed_page = get_page_data_from_file_path(
+                markdown_file, strip_header=strip_header
+            )
             processed_page.parent_title = parent_page_title
             processed_pages.append(processed_page)
 
@@ -152,14 +155,14 @@ def get_pages_from_directory(
     return processed_pages
 
 
-def get_page_data_from_file_path(file_path: Path) -> Page:
+def get_page_data_from_file_path(file_path: Path, strip_header: bool = False) -> Page:
     if not isinstance(file_path, Path):
         file_path = Path(file_path)
 
     with open(file_path) as file_handle:
         markdown_lines = file_handle.readlines()
 
-    page = get_page_data_from_lines(markdown_lines)
+    page = get_page_data_from_lines(markdown_lines, strip_header=strip_header)
 
     if not page.title:
         page.title = file_path.stem
@@ -169,20 +172,22 @@ def get_page_data_from_file_path(file_path: Path) -> Page:
     return page
 
 
-def get_page_data_from_lines(markdown_lines: List[str]) -> Page:
+def get_page_data_from_lines(
+    markdown_lines: List[str], strip_header: bool = False
+) -> Page:
     frontmatter = get_document_frontmatter(markdown_lines)
     if "frontmatter_end_line" in frontmatter:
         markdown_lines = markdown_lines[frontmatter["frontmatter_end_line"] :]
 
-    page = parse_page(markdown_lines)
+    page = parse_page(markdown_lines, strip_header=strip_header)
 
     if "title" in frontmatter:
         page.title = frontmatter["title"]
     return page
 
 
-def parse_page(markdown_lines: List[str]) -> Page:
-    renderer = ConfluenceRenderer(use_xhtml=True)
+def parse_page(markdown_lines: List[str], strip_header: bool = False) -> Page:
+    renderer = ConfluenceRenderer(use_xhtml=True, strip_header=strip_header)
     confluence_mistune = mistune.Markdown(renderer=renderer)
     confluence_content = confluence_mistune("".join(markdown_lines))