+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+- Add feature to only update things that have changed by storing the SHA1 in the update message
 - Make attachment updates work on both Confluence Server and Confluence Cloud
 
 ## 1.0.4 - 2021-11-25